Share solver for all snapshots
Share the same solver for all snapshots. This is accomplished by pushing
to the solver for each snapshot, so that we can pop it when we need to
backtrack.

Most of the time, snapshots will depend on each other, so we can skip
recreating the solver for each snapshot when at most only a few
constraints are added for every snapshot.