Fix FlowSnapshot#commonPredecessor
Fix FlowSnapshot#commonPredecessor to correctly determine the common
predecessor. This allows the FlowSolver to skip having to backtrack.
This fix decreased the SatisfiabilityPush and SatisfiabilityPop
statistics from 1483-1431 to 157-105.

Furthermore, we also switched the order of the assertions in the branch
instruction. Assuming the then-branch will occur more often than the
else-branch. This means that the FlowSolver is already configured on
the then-branch, which is also the next snapshot to be processed -
meaning it doesn't need to backtrack. This decreased the
SatisfiabilityPush and SatisfiabilityPop statistics to 107-55.